[TECH-521] Move build logic back to top level
branch: feature/TECH-521-simplified-remote-artifact-caching
